mididrone_splitter: New dispatcher: ChannelDispatcher
Add the ChannelDispatcher, a dispatcher which maps every channel to a
musician and tries to assign notes to that musician before trying any
other.
The result is that one musician has a greater chance to play the same
voice for the whole sequence.
Also created the Dispatcher base abstract class from which all the
dispatcher implementations should derive.