New dispatcher: PriorityMapDispatcher
Add new priority map dispatcher, which dispatches notes to musicians
according to the priority and set of channels assigned to them by the
user.

This allows more complex mappings, to get a specific set of channels to
play only on some musicians, or make a musician play a solo of one or
more channel, excluding all others.